+++ test/tutorial-2.dart
@@ -45,9 +45,8 @@ Future<Response> _myExceptionHandler(
   String message;
 
   if (ex is NotFoundException) {
-    status = (ex.found == NotFoundException.foundNothing)
-        ? HttpStatus.methodNotAllowed
-        : HttpStatus.notFound;
+    status =
+        ex.resourceExists ? HttpStatus.methodNotAllowed : HttpStatus.notFound;
     message = 'Sorry, the page you were looking for could not be found.';
   } else {
     status = HttpStatus.internalServerError;